updated info for VIP integration
