+++ config.txt
@@ -1,0 +1,11 @@
+# Variable names and section titles should stay the same
+# Put this file in your HomeFolder/.anima/config.txt
+# Make the anima variable point to your Anima public build
+# Make the extra-data-root point to the data folder of Anima-Scripts
+# The last folder separator for each path is crucial, do not forget them
+# Use full paths, nothing relative or using tildes 
+
+[anima-scripts]
+anima = /Segmentation-Pipeline/Anima-Binaries-4.0.1/
+anima-scripts-public-root = /Segmentation-Pipeline/Anima-Scripts-Public/
+extra-data-root = /Segmentation-Pipeline/Anima-Scripts-Data-Public/
