+++ requirements.txt
@@ -3,6 +3,7 @@ chardet==4.0.0
 click==7.1.2
 configparser==5.0.2
 Deprecated==1.2.12
+dill==0.3.4
 docker-pycreds==0.4.0
 gitdb==4.0.7
 GitPython==3.1.14
